+++ package.json
@@ -16,6 +16,7 @@
     "@types/react": "^18.2.14",
     "@types/react-dom": "^18.2.6",
     "react": "^18.2.0",
+    "react-daum-postcode": "^3.1.3",
     "react-dom": "^18.2.0",
     "react-material-ui-carousel": "^3.4.2",
     "react-router-dom": "^6.15.0",
